fix: load BTC address in ckBTC receive modal
